personal_knowledge_assistant/requirements.txt
flask
flask-cors
sentence-transformers
faiss-cpu
numpy
tqdm
